(Un)Lock properly and add methods for acquiring data
Only hold the lock when parameters are being written to. Add
parameters to allow acquisitions longer than the Xspress3 buffer
size. Add grabFrame to wrap up readFrame, SCA publishing and parameter
updates during acquisition. Add checkQueue to look for specific events
on the queue. Add doALap to read an Xspress3 buffers worth of frames
rounded down to a whole number of rows. Add startAcquisition to deal
with set up, running of sufficient laps and wrap up at the end of an
acquisition. Wait for start events in mainLoop and start acquisitions
when requested.