Keep track of frame no; check for stop every frame
There was a race condition when relying on NDArrayCounter to keep
track of the frame number that caused errors at high frame rates so
keep track locally. The queue was only being checked when there wasn't
a new frame so it was impossible to stop the acquisition at high frame
rates.