Remove maxDriverFrames
Max driver frames no longer applies as the number of frames that can
be collected is unbounded now. The number of frames that the hardware
can buffer currently sets a limit on the number of frames that can be
captured before a new "acquisition" must be started; this means that
the time between frames will, most likely, be uneven between the end
of one "acquisition" and the next.