Update xspress3Channel.template
Change R macro to S so that R can be rebound. Add INDEX macro as the
arrays are 0 indexed and there's no subtraction operator in msi. Remove
the SCA and ROI substitutions and make SCAs as an NDAttribute
plugin. Add plugins for channel visualisation. Remove Attr templates
from channel template.